backend/requirements.txt
numpy==1.26.4
Cython==3.0.12
duckdb==1.2.2
duckdb_engine==0.17.0
fastapi==0.115.12
fastapi-cli==0.0.7
pandas==2.2.3
requests==2.32.3
scikit-surprise==1.1.4
SQLAlchemy==2.0.40
uvicorn==0.34.2

